requirements.txt
flask==2.3.2
flask-cors==3.0.10
torch>=1.8.1
torchvision==0.18.0
Pillow==8.2.0
requests==2.32.3
gunicorn==20.1.0
isort==5.13.2
pytest==6.2.4
flake8==6.0.0
flake8-isort==5.0.3
black==22.3.0
types-requests
numpy
mypy==1.3.0
mypy-extensions==1.0.0
urllib3==2.2.2
sqlalchemy
spacy
psycopg2
transformers
pandas
scikit-learn
google-generativeai
streamlit
pinecone-client
openai
python-dotenv==0.19.2
semantic-router